requirements.txt
asgiref==3.3.1
Django==3.1.7
django-filter==2.4.0
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
yapf==0.30.0
